pyproject.toml
[tool.poetry]
name = "digg"
version = "0.1.0"
description = ""
authors = ["Caio Dadauto <caiodadauto@gmail.com>"]
readme = "README.md"
build = 'build.py'

[tool.poetry.dependencies]
python = "~3.10"
conda-lock = "^1.4.0"
pandas = "^1.5.2"
seaborn = "^0.12.2"
networkx = "^2.8.8"
mlflow = "^2.2.1"
hydra-core = "^1.3.2"
tqdm = "^4.65.0"
prettytable = "^3.6.0"
click = "^8.1.3"
joblib = "^1.2.0"
pyemd = "^1.0.0"
torch = "1.13.0"
# pyside6 = "6.4.2"
numpy = "1.23.5"

[tool.poetry.scripts]
digg = "digg.cli.app:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
